refactor: small code cleanup
