+++ src/Pages/ProgramsPage/ProgramsPage.tsx
@@ -2,7 +2,7 @@ import { memo, useCallback, useEffect, useState } from 'react'
 
 import { useNavigate, useParams } from 'react-router-dom'
 
-import { Delete, FileOpen, Save, SaveAlt } from '@mui/icons-material'
+import { Delete, Save, SaveAlt } from '@mui/icons-material'
 import {
   IconButton,
   List,
@@ -13,8 +13,13 @@ import {
   Tooltip,
 } from '@mui/material'
 
-import { AdminPageWrapper, LoadingAnimation, TreeList } from '../../components'
-import { useProgram, useSupabase } from '../../hooks'
+import {
+  AdminPageWrapper,
+  LoadingAnimation,
+  TreeList,
+  LoadLocalProgramButton,
+} from '../../components'
+import { useSupabase } from '../../hooks'
 import { useAdminStore } from '../../stores'
 import { DbProgram } from '../../types'
 import { saveProgramAsJson } from '../../utils'
@@ -29,7 +34,7 @@ const StyledProgramsListContainer = styled(List)`
 export const ProgramsPageRaw = () => {
   const navigate = useNavigate()
   const { programId } = useParams()
-  const { loadJsonProgram } = useProgram()
+
   const program = useAdminStore((s) => s.program)
 
   const {
@@ -67,12 +72,6 @@ export const ProgramsPageRaw = () => {
     navigate(`/admin/programs`)
   }, [deleteProgram, navigate, program])
 
-  const handleImportJsonProgram = useCallback(
-    (e: any) => {
-      loadJsonProgram('/' + e.target.files[0].name)
-    },
-    [loadJsonProgram]
-  )
   const handleSaveProgramAsJson = useCallback(() => {
     if (!program) return
     saveProgramAsJson(program)
@@ -82,19 +81,7 @@ export const ProgramsPageRaw = () => {
     return (
       <AdminPageWrapper
         topNavHeader="programs"
-        topNavActions={
-          <>
-            <IconButton component="label">
-              <input
-                hidden
-                accept="application/json"
-                type="file"
-                onChange={handleImportJsonProgram}
-              />
-              <FileOpen />
-            </IconButton>
-          </>
-        }
+        topNavActions={<LoadLocalProgramButton />}
       >
         {error && <h3>Something went Wrong: {error}</h3>}
         {!userIsLoggedIn ? (
