+++ src/components/index.ts
@@ -10,3 +10,4 @@ export * from './ScreenFormPopup'
 export * from './AdminPageWrapper'
 export * from './LoadingAnimation'
 export * from './LoginPopup'
+export * from './LoadLocalProgramButton'
